+++ RELEASE_NOTES.md
@@ -1,8 +1,25 @@
 **2026.24.2**
 
-**Json Payload Information**
-* Added "Json" Payload Tab to Groups
-* Added "Json" Payload Tab to Worlds
-* Added "Json" Payload Tab to Avatars
-* Added "Json" Payload Tab to User Profiles
-* Added "Json" Payload Tab to Own Profile
+**JSON Payload Information**
+* Added a **JSON Payload** tab to Groups.
+* Added a **JSON Payload** tab to Worlds.
+* Added a **JSON Payload** tab to Avatars.
+* Added a **JSON Payload** tab to User Profiles.
+* Added a **JSON Payload** tab to Own Profile.
+
+**Group Modal**
+* Moved the **Change Banner** pen next to the **Share** button.
+
+**Own Profile Modal**
+* Moved the **Change Banner** pen next to the **Share** button.
+* Added a **Share** button to Own Profile.
+
+**Avatar Modal**
+* Added a **Change Banner** pen, so you can now change the image of your avatar inside VRCN.
+* Added a **Gallery** tab inside the Avatar Modal, so you can view gallery images of an avatar.
+* Added an **Upload Image** button to upload images to your own avatar gallery.
+* Added image compression for images larger than 5 MB.
+* Added an image crop tool for images larger than 2000x2000 pixels.
+
+**i18n**
+* Added missing translation keys for Avatar Modals